Removed duplicated method that was not needed and refactored parts of userAccount for spelling/clarity. #6
